requirements.txt
langchain
langchain-openai
matplotlib
numpy
pandas
plotly
python-dotenv
scikit-learn
scipy
seaborn
statsmodels
streamlit
wordcloud